+++ backend/src/main/resources/application.properties
@@ -3,6 +3,6 @@ spring.datasource.url=jdbc:postgresql://localhost:5432/homet_test_voice
 spring.datasource.username=user
 spring.datasource.password=
 spring.datasource.driver-class-name=org.postgresql.Driver
-spring.jpa.hibernate.ddl-auto=update
+spring.jpa.hibernate.ddl-auto=create
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
